Backend_FastAPI/requirements.txt
fastapi>=0.95.0,<0.100.0
uvicorn>=0.20.0,<0.24.0
python-multipart>=0.0.5
python-dotenv>=0.21.0
pillow>=9.3.0
numpy>=1.21.0,<1.25.0
tensorflow>=2.10.0,<2.14.0
pydantic>=1.10.0,<2.0.0
python-jose>=3.3.0
passlib>=1.7.4